Lab 6 Code and final paper
Missing raw data, and thats needed for the parser to work. 
some people had issues with the mapping, be sure to put your code in the same folder as the raw data, just to be safe. also, the folder with code+raw data should be in the expected MATLAB mapping folders.